# src/main/resources/openapi/Accounts.yaml
openapi: 3.0.3
info:
  title: Accounts
  description: Manage the accounts
  version: 1.0.0
servers:
  - url: http://localhost/accounts
paths:
  /createAccount:
    post:
      summary: Create an Account and registered it to an user
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '200':
          $ref: '#/components/schemas/AccountResponse'
        '503':
          $ref: '#/components/schemas/Error'

  /addCreditsToAccount:
    post:
      summary: Add Credits To Account.
      description: Optional extended description in CommonMark or HTML.
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCredits'
      responses:
        '200': # status code
          description: Operation success
        '404':
          $ref: '#/components/responses/NotFound'

  /associateUserToAccount:
    post:
      summary: Associate User With An Account
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociateUser'
      responses:
        '200':
          description: Operation success'
        '404':
          $ref: '#/components/responses/NotFound'

  /linkAccountToMercadoPago:
    post:
      summary: Link account to mercado pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkAccountToMercadoPago'
      responses:
        '200':
          description: Operation success
        '404':
          $ref: '#/components/responses/NotFound'
  /buyCreditsMercadoPago:
    post:
      summary: Buy credits in Mercado Pago
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyCreditsMercadoPago'
      responses:
        '200':
          description: Operation success
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    Error:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    Account:
      type: object
      properties:
        balance:
          type: number
          format: double
        active:
          type: boolean
    CreateAccount:
      type: object
      required:
        - userId
        - account
      properties:
        userId:
          type: integer
          format: int32
        account:
          $ref: '#/components/schemas/Account'
    AddCredits:
      type: object
      required:
        - accountId
        - credits
      properties:
        accountId:
          type: integer
          format: int32
        credits:
          type: number
          format: double
          minimum: 0.0
          exclusiveMinimum: true
    AssociateUser:
      type: object
      required:
        - userId
        - accountId
      properties:
        accountId:
          type: integer
          format: int32
        userId:
          type: integer
          format: int32
    LinkAccountToMercadoPago:
      type: object
      required:
        - accountId
        - mercadoPagoId
      properties:
        accountId:
          type: integer
          format: int32
        mercadoPagoId:
          type: integer
          format: int32
    BuyCreditsMercadoPago:
      type: object
      required:
        - accountId
        - credits
      properties:
        accountId:
          type: integer
          format: int32
        credits:
          type: number
          format: double
          minimum: 0.0
          exclusiveMinimum: true
    AccountResponse:
      type: object
      properties:
        accountId:
          type: integer
          format: int32
  responses:
    Error:
      description: Error accessing the service consistent with http status code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidInput:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Unable to find resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccountResponse:
      description: Create account Success response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AccountResponse'